+++ common/utils.ts
@@ -1,4 +1,4 @@
-type Inputs = Record<string, any>
+export type Inputs = Record<string, any>
 
 export const validateInputs = (inputs: Inputs): boolean => {
   for (const key in inputs) {
@@ -7,7 +7,7 @@ export const validateInputs = (inputs: Inputs): boolean => {
     }
 
     if (typeof inputs[key] === 'object') {
-      if (!validateInputs(inputs[key])) {
+      if (!validateInputs(inputs[key] as Inputs)) {
         return false
       }
     }
