--- node/events/index.ts
@@ -1,5 +1,0 @@
-export * from './cancelOrder'
-export * from './eventsErrorHandler'
-export * from './compareOrder'
-export * from './updateGlovoOrderStatus'
-export * from './validateEventSettings'
