+++ node/middlewares/events/index.ts
@@ -1,0 +1,13 @@
+import { compareOrder } from './compareOrder'
+import { eventsErrorHandler } from './eventsErrorHandler'
+import { updateGlovoOrderStatus } from './updateGlovoOrderStatus'
+import { validateEventSettings } from './validateEventSettings'
+
+export const events = {
+  updateGlovoOrder: [
+    eventsErrorHandler,
+    validateEventSettings,
+    compareOrder,
+    updateGlovoOrderStatus,
+  ],
+}
