+++ node/middlewares/events/validateEventSettings.ts
@@ -1,6 +1,6 @@
 import { UserInputError } from '@vtex/api'
 
-import { APP_SETTINGS, GLOVO } from '../constants'
+import { APP_SETTINGS, GLOVO } from '../../constants'
 
 export async function validateEventSettings(
   ctx: StatusChangeContext,
