--- node/middlewares/index.ts
@@ -1,15 +1,0 @@
-export * from './authorizeOrder'
-export * from './createOrder'
-export * from './errorHandler'
-export * from './getGlovoMenu'
-export * from './glovoMenuCompleteUpdate'
-export * from './glovoMenuPartialUpdate'
-export * from './glovoProductUpdate'
-export * from './saveGlovoMenu'
-export * from './saveOrderRecord'
-export * from './sendReponse'
-export * from './simulateOrder'
-export * from './storeMenuUpdate'
-export * from './validateSettings'
-export * from './validateGlovoToken'
-export * from './getGlovoMenuByStore'
