+++ node/middlewares/routes/simulateOrder.ts
@@ -5,7 +5,7 @@ import {
   convertGlovoProductToItems,
   createSimulationPayload,
   getStoreInfoFormGlovoStoreId,
-} from '../utils'
+} from '../../utils'
 
 export async function simulateOrder(ctx: Context, next: () => Promise<void>) {
   const {
@@ -15,6 +15,9 @@ export async function simulateOrder(ctx: Context, next: () => Promise<void>) {
   } = ctx
 
   const glovoOrder: GlovoOrder = await json(ctx.req)
+
+  // eslint-disable-next-line no-console
+  console.log({ glovoOrder })
 
   ctx.state.glovoOrder = glovoOrder
 
@@ -70,6 +73,9 @@ export async function simulateOrder(ctx: Context, next: () => Promise<void>) {
       message: `Simulation for order ${glovoOrder.order_id}`,
       simulation,
     })
+
+    // eslint-disable-next-line no-console
+    console.log({ simulation })
 
     ctx.state.orderSimulation = simulation
     ctx.state.storeInfo = storeInfo
