+++ node/typings/general.d.ts
@@ -202,3 +202,63 @@ interface VTEXAuthorizedMarketplaceOrder {
     ]
   }
 }
+
+interface SimulationItem {
+  id: string
+  quantity: number
+  sellerId: string
+}
+
+interface SimulatedItem {
+  price: number
+  available: boolean
+}
+
+interface CreateSimulationArgs {
+  items: PayloadItem[]
+  postalCode?: string
+  country?: string
+  affiliateId: string
+  salesChannel: string
+}
+
+interface OrderRecord {
+  orderId: string
+  glovoOrder: GlovoOrder
+  invoiced: any | null
+  hasChanged: boolean
+  createdAt?: number
+  startHandlingAt?: string
+  invoicedAt?: number
+}
+
+interface ProductRecord {
+  id: string
+  price?: number
+  available?: boolean
+}
+
+interface StoreMenuRecord {
+  items: {
+    [id: string]: ProductRecord
+  }
+  lastUpdated: string
+}
+
+interface GlovoMenu {
+  [key: string]: boolean
+}
+
+interface StoreMenuUpdates {
+  current: MenuUpdatesItem
+  previous?: MenuUpdatesItem
+}
+
+interface MenuUpdatesItem {
+  responseId: string | null
+  createdAt: number
+  storeId: string
+  storeName: string
+  glovoStoreId: string
+  items: ProductRecord[]
+}
